+++ src/Type/Definition/FieldDefinition.php
@@ -11,6 +11,7 @@ use GraphQL\Utils\Utils;
 
 use function is_array;
 use function is_callable;
+use function is_iterable;
 use function is_string;
 use function sprintf;
 
@@ -86,9 +87,9 @@ class FieldDefinition
             $fields = $fields();
         }
 
-        if (! is_array($fields)) {
+        if (! is_iterable($fields)) {
             throw new InvariantViolation(
-                "{$type->name} fields must be an array or a callable which returns such an array."
+                "{$type->name} fields must be an iterable or a callable which returns such an iterable."
             );
         }